integration --create: fix option parsing on dash
With dash, "shift" when $# = 0 causes a fatal error, killing the shell
instead of simply returning non-zero from the builtin command.  This
causes problems with the code in integration_create, which does:

	base=${1-master}
        shift

Obviously if the default base is to be used then $# = 0 at this point,
which causes the following "shift" to fail.

Fix this by avoiding shifting away this parameter and changing the
"correct usage" test to detect $# -le 1.

Signed-off-by: John Keeping <john@keeping.me.uk>