Allow --rebuild to be specified with --edit
This re-writes the option parsing code so that --create, --edit and
--rebuild are treated as independent flags, which can be combined in any
combination.  When more than one is specified, they are applied in the
order "create branch", "edit instruction sheet", "rebuild branch".

This re-write removes the "base cannot be specified if --create and
--edit are used".  Positional parameters are now handled correctly for
all combinations - only --create uses more than one argument and the
first argument has the same meaning as for --edit and --rebuild.

Signed-off-by: John Keeping <john@keeping.me.uk>