Die if run_integration fails
Since we pipe data into run_integration, it runs in a subshell and hence
can die without causing git-integration itself to fail.

Currently this isn't a problem since the call to run_integration is the
last thing to happen in integration_rebuild, which is in turn only ever
called at the end of git-integration.  But if another command that get
sequenced after rebuild is added, we will end up running that when we
shouldn't or hiding the exit code from integration_rebuild.

Signed-off-by: John Keeping <john@keeping.me.uk>