Silent "no such file" when integration is not in progress
When 'git integration' is invoked with --continue or --abort but there
is no integration in progress, we attempt to get the contents of a
status file that does not exist.  This causes cat to print a "no such
file" error which presents the user with irrelevant implementation
details.  Suppress this.

Signed-off-by: John Keeping <john@keeping.me.uk>