Simplify usage examples
This reduces the different types of usage to three examples and makes
the man page content and that produced by 'git integration -h' the same.

We now have:

	git integration --create <name> [<base>]
	git integration [<action>...] [<branch>]
	git integration --continue | --abort

This is an improvement on the previous version which was missing several
options and would grow unwieldy if all of the available options are
added.

The three cases given above simply identify the one case where the
branch name is required (and base branch is optional), the cases where a
branch name is optional and any number of actions may be specified and
the cases where only a single option may be provided.

Signed-off-by: John Keeping <john@keeping.me.uk>