Allow user-specified merge options
When adding a git-gui specific change to my git.git integration branch,
the merge needs the 'subtree' strategy to be used since git-gui is
developed independently and periodically merged into git.git as a
subtree.

In order to accommodate this, allow options to be specified after the
ref name in a 'merge' instruction and pass these options directly to
git-merge.  It would be nice to permit the ref and options to be given
in any order but that would require git-integration to know about all
options to git-merge (in order to handle unstuck arguments to options)
and I don't think that's necessary at the moment.

Signed-off-by: John Keeping <john@keeping.me.uk>