Add a "fixup" instruction
This is useful when a logical fix needs to be added as part of the
"evil" merge which brings a branch into the integration branch.  The
"fixup" instruction cherry-picks a commit, squashing it into the current
tip commit of the integration branch - i.e. the last change that was
made.

The status output for this is interesting, since Git does not provide an
easy way to see if a particular change has been squashed.  This patch
simply tries to apply the patch for the change to the index (using
"git apply --cached") and assumes that if that operation fails, then the
patch is already squashed in.  This isn't particularly reliable since
git-apply in this mode will not attempt any form of merge when applying
the patch but might work well enough in practice and is at least nice
and simple.

Signed-off-by: John Keeping <john@keeping.me.uk>