Add a "fixup" instruction
This is useful when a logical fix needs to be added as part of the
"evil" merge which brings a branch into the integration branch.  The
"fixup" instruction cherry-picks a commit, squashing it into the current
tip commit of the integration branch - i.e. the last change that was
made.

The status output for this is interesting, since Git does not provide an
easy way to see if a particular change has been squashed.  I have
adopted an approach which identifies the changes added by the fixup and
checks to see that they were indeed added at some point in the history
of each of the files involved.  This probably needs refining to deal
with fixups that only remove code, but it's not obvious to me at the
moment precisely how this can be detected, since there is no guarantee
that the removed lines have actually existed in the Git history - they
may have been added by the commit into which the fixup has been
squashed; this isn't possible if the fixup has created an evil merge but
may apply if the fixup is on top of a cherry-picked commit or even
another fixup.

Signed-off-by: John Keeping <john@keeping.me.uk>