status: include SHA-1 of merge to integration branch
In the status output for a "merge" command, add the SHA-1 of the commit
in the integration branch which represents this command, if we know it.

The implementation assumes that there will only be one commit on the
integration branch which has the specified commit as its last parent
(i.e. its second parent for non-octopus merges).

Signed-off-by: John Keeping <john@keeping.me.uk>