Don't create empty commits in the instruction ref
If the user asks to edit the instruction sheet but does not actually
change anything, then git-integration currently creates an empty commit
on the instruction sheet branch.

This is unnecessary and makes it harder to find genuine changes to the
instruction sheet, so detect this situation and do not create a new
commit if it would be empty.

Signed-off-by: John Keeping <john@keeping.me.uk>