create: don't call "test" with invalid arguments
When we first create a new integration branch, there is no parent
commit, so $parent_tree is the empty string.  This means that "test" is
invoked as "test <new_tree> =" which is invalid.

There is no correctness impact here because the error is the same as the
comparison failing, but it does produce a scary error message from the
shell.  Add DQs around the variables in the call to "test" to suppress
this.

Signed-off-by: John Keeping <john@keeping.me.uk>