Add new "--prefix=..." option
This allows the user to specify a prefix to be prepended to all refs in
an instruction sheet when building or inspecting an integration branch.

The intended use case is for secondary users using someone else's
instruction sheet, who may have the branches under a different ref
namespace; for example, the refs that are under refs/heads/ for a
maintainer may be under refs/remotes/origin/ for users.

This option can also be configured using the "integration.prefix"
configuration key.

The Bash completion script suggests the first component of each ref
under the refs/remotes/ hierarchy.

Signed-off-by: John Keeping <john@keeping.me.uk>