Improve reflog action descriptions
git-integration currently relies on setup_reflog_action prefixing all
changes with "integration: ", but this does not show what is happening
when refs are updated in all cases.  Improve this by passing an explicit
"-m" argument to "git update-ref" and "git symbolic-ref" and overriding
$GIT_REFLOG_ACTION when calling "git checkout" or "git reset" in order
to provide a better description of what's going on.

Signed-off-by: John Keeping <john@keeping.me.uk>