Add an "exec" instruction
This allows a command to be executed as part of the integration rebuild
process, which may be useful, for example, to execute a script to update
submodules to their own suitable integration versions.

This instruction is disabled by default so as to avoid accidentally
running a malicious command in an instruction sheet acquired from
elsewhere.  Configuration keys can be used to enable it.

Signed-off-by: John Keeping <john@keeping.me.uk>