Allow --cat and --status to be performed while rebuilding
There is no reason read-only operations should not be permitted while a
rebuild is in-progress.  "--status" might do some weird things if the
branch changes underneath it, but the primary use-case for this is
inspecting the previous state of the branch while paused due to a
conflict during the rebuild.

Signed-off-by: John Keeping <john@keeping.me.uk>