Fix --cat and --status while rebuilding
In commit 165e5b2 (Allow --cat and --status to be performed while
rebuilding, 2014-01-11) we allowed --cat and --status to be used while a
rebuild is in-progress but this has never worked unless --no-rebuild is
also specified.  I think that the commit allowing this must originally
have been written before commit 872a399 (Add "integration.autorebuild"
configuration variable, 2013-03-30) was merged.

Let's add a test to make sure this doesn't break again in the future.

Signed-off-by: John Keeping <john@keeping.me.uk>