Preserve autocontinue state after stopping
If a rebuild is started with --[no-]autocontinue, then after the first
merge failure the state reverts to whatever is specified in the git
config.  Save the autocontinue state and restore it during
"git integration --continue" so that the value given with --rebuild is
used for the entire integration run.

Signed-off-by: John Keeping <john@keeping.me.uk>